+++ src/main/kotlin/com/atlassian/performance/tools/jirasoftwareactions/api/boards/Board.kt
@@ -1,0 +1,7 @@
+package com.atlassian.performance.tools.jirasoftwareactions.api.boards
+
+interface Board {
+    val id: String
+    val issuesOnBoard: Int?
+    fun isWorthVisiting(): Boolean
+}