+++ CHANGELOG.md
@@ -26,6 +26,13 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ## [Unreleased]
 [Unreleased]: https://bitbucket.org/atlassian/jira-software-actions/branches/compare/master%0Drelease-1.1.0
 
+### Added
+- New Board class
+- Memory for boards
+
+### Deprecated
+- AgileBoardIdMemory
+
 ## [1.1.0] - 2018-10-15
 [1.1.0]: https://bitbucket.org/atlassian/jira-software-actions/branches/compare/release-1.1.0%0Drelease-1.0.0